Remove the peephole optimizer.
The optimizer is quite brittle. It relies heavily on hints and the
compiler's code generation being fairly uniform. Any kind of unexpected
pattern can introduce bugs, as it does not perform a proper use/def
analysis or respect control flow.

It also doesn't really matter. Half of what it does is insert macro ops.
Macro ops are pointless given that we have a JIT, and in fact macro ops
have been removed from upstream Pawn. The other optimizations are
relatively minor, and could be brought back with a proper optimizer in
the VM or in the local compiler.

Since the optimizer has been disabled for expressions in 1.11, let's
finally just remove the code.