Move function parsing to the AST pipeline.
This change was very complicated due to the sheer amount of stuff that
delicately depended on the parsing/semantic/codegen interleaving in
newfunc().

One of the biggest changes is that temporary heap allocations no longer
get collected into the nearby block scope. Instead, they go into a
statement-level temporary scope, similar to how temporary allocations
work in C++. This makes it much easier for us to emit verifiable code,
and ensures that allocations are freed closer to their point of use. The
downside is that temporary scopes must be inserted very carefully. Any
time branching might happen within an expression, there must be a
temporary scope. Otherwise, the deallocation might not be dominated by
the allocation in the control-flow graph.

In the future, it would be much simpler to introduce a new opcode that
pins and resets the heap.