Fix data flow bug in the method verifier.
The method verifier attempts to track the heap and "heap tracker" in the
same list, but this is not really correct. It's two separate lists.
Treating them as one loses the ability to verify joins. For example:

    block1:
        tracker.push.c 0
	jump block3
    block2:
        heap 8
        tracker.push.c 8
        jump block3
    block3:
	tracker.pop.setheap

This flow is valid because at the join point, the heap can be cleaned up
regardless of which path was taken. But in the existing verifier code,
the heap appears imbalanced.

We fix this by splitting it into lists. If all paths through a block
have the same static, determinate value for a tracker entry, that entry
is allowed to flow through. Otherwise, the join point will track the
value as indeterminate instead.

Because of the way we iterate blocks, any join point will inherit its
first predecessor's state. Thus we need some complicated handling to
"merge" the state when we analyze the next predecessor. This code is
fiddly, but not complicated: the goal is to make a list of determinate
allocations look like a list of unknown allocations.

Additionally, this fixes a long-standing verification bug where we did
not correctly analyze the "exit" state of blocks compared to the "entry"
state of their dominated blocks. We now track entry state separately
from exit state, which allows us to analyze backedges properly.

Test: plugin corpus