Fix ternary operator heap tracking.
This bug is so obscure it was only caught by a single file in the forums
corpus. But, it is simple to explain:

    int blah[3]
    blah = Stuff() ? {0, 0, 0} : Something();

The first "arm" of the ternary returns a static array. The other arm
returns a dynamic array. This caused us to emit a heap tracking opcode
on the second arm, but not the first. The imbalance in the tracking
state causes a verification failure.

The fix is to generate the code once, and throw it away and fix it if
we detect this scenario. There is surely a cleverer fix but it can wait
until we have moved entirely to the AST pipeline.

Test: reduced test case