Refactor how source files are read.
This is a big simplification that removes libpawnc. Source files are now
read into an std::string and tracked as a C++ object. Separate tracking
of file names is gone.

This removes the "#file" preprocessor command. It also removes the
ability to merge multiple source files, eg:

    spcomp a.sp b.sp c.sp

This usage has never been encouraged and its use would be unexpected.

If we get requests to add it back, the lexer should instead use a stack
of SourceFiles rather than merging all the source files into one big
one. This will be easier to implement when reparsing is removed.