Use SymbolScope for globals.
This removes sp_symhash, and replaces all lookups with SymbolScope. This
contains a simple map of atoms (names) to symbol lists. The list form
will be used for operators in a successive patch. For now, it's used for
statics (explained later).

This patch is a bit rough and requires a bunch of invasive changes to
symbol lookup. Ideally, all lookups and all symbol entry would go
through a normal scope chain. But, Pawn being what it is, we cannot do
this right away. The reason is reparsing.

Pawn relies heavily on retaining certain symbols across reparses. This
mechanism is difficult to maintain if we split the global scope into two
levels (static level and global level). We have no easy way to rebuilding
file-level scopes across reparses, so instead, we allow duplicate global
symbols as long as their scope is technically different (eg, static and
non static, or statics in different files).

As a result, we need to pass the parsed file number around to all symbol
lookups, until we can eliminate reparsing.

This patch also eliminates a bunch of complexity in add_constant. We no
longer allow redefining constants under bizarre conditions. Instead,
it's always illegal to reintroduce a name in the same scope.

This patch also converts a lot more C-style strings into Atoms, which
should be a minor improvement to compilation speed.

Also, a long-standing issue with assemble() has been fixed. It now
embeds symbol pointers instead of names, to avoid erroneous collisions.

Finally, CompileContext has been introduced. This is where scvars will
start moving to, and dropping scvars is a necessity to making the
compiler parallelizable.