Restrict when enum names can be used as a size value.
Although we now allow enum names as sizes again, this feature is getting
restricted to only work on array declarations. In any other context it
will error.

This does break a handful of plugins, and those plugins are often doing
something erroneous (like "return Action"). That is the intent behind
this change: it's very difficult to distinguish between valid uses and
typos. Other uses appearing valid actually have off-by-one errors.

As a compromise, we now have a magic property on enum names, "__next".
This property evaluates to the next value in the enumeration. Eg,

    enum Stuff { A, B, C };

The expression `Stuff` is no longer valid, except for array dimensions.
To access the array size, either `sizeof` or `Stuff.__next` can be used.