Parse include directives as statements.
This is a huge hack, but it will make it much easier to properly clean
up the implementation of scopes. The problem is that #include statements
can appear anywhere, eg:

     public void
     #include "blah.sp"
     () {}

In the current system, static variables are globals that are tied to a
specific file. This requires attaching a file indicator to every name
lookup. It would be much cleaner to place statics into a dedicated
scope.

In the above example, "blah.sp" introduces a new static scope while in
the middle of parsing a function. This is extremely difficult to deal
with and exposes a bunch of crazy edge cases.

The effect of this change is that #include is now legal ONLY as a
top-level statement. It acts much like a declaration now. Due to the
lexer being very tricky, this works by injecting temporary tokens into
the token stream. These tokens are used by the parser to open the file
and reset the lexer.

Thus the above code is now illegal, as are lots of other #include
patterns that would be legal in C.