Invert the code generator's AST walk.
This moves codegen out of the AST itself and into a separate generator
class, which manually walks the AST. The primary reason is to improve
encapsulation. There are many nodes that don't emit any code and don't
neatly slot into the previous design. The new design makes it easier to
have separate code generators. It also makes it easier to remove the
text-based emitter, since attaching all the emitter code to input
variables would be really cumbersome.

Currently the generator uses a big switch statement to avoid the cost of
two virtual dispatches.

A similar refactoring for SemaContext will likely happen as well.