Remove the text-based assembler.
This improves compilation performance by about 10%, and eliminates a
huge pile of legacy code. Previously, the code generator emitted a text
stream containing assembly instructions. This was then either written to
disk (if -a was given on the command-line), or re-parsed to emit actual
instructions.

This system did have certain advantages, namely, it's easy to debug and
it's easy to rewind the assembly stream. But it had many disadvantages.
It's slow, it's a lot of code, it's difficult to extend, and it required
a lot of hacks to make symbol references and the data section work.

The new code generator emits a binary stream without the text processing
step. It is much more straightforward. We already have smxviewer to
inspect files, so debuggability is not lost.

The new generator uses code from the experimental branch, namely, Label.
Labels greatly simplify backpatching by embedding linked lists within
the codegen stream. This unfortunately breaks the ability to "rewind"
the codegen process (which was called "staging" mode). Staging mode was
used to make sure ternary operators were emitted correctly.  For now, we
emit some relatively gross but harmless instructions to paper over the
problem. Once the analysis phase has been refactored we'll be able to
annotate expressions before codegen begins, and provide a priori
information to eliminate the need for staging hacks.

A new option, --no-verify, has been added as well. When verification
fails it's useful to still emit a .smx file so it can be debugged in
smxviewer.

It's also worth noting that this patch removes a large number of global
variables. sc_labnum, glb_declared, the text stream (gAsmBuffer),
curseg, and the particularly horrible code_idx are now gone.