Use visitor pattern (kinda) for semantic analysis.
Following the refactoring of CodeGenerator, this moves semantic analysis
into its own class and moves analysis out of ParseNode. This gives us a
lot more flexibility and means we have a consistent semantic object
across checks.

A similar refactoring could be done for binding/name resolution. For now
there's no real reason to.

This patch is also the start of moving away from Mozilla syntax.
Function signatures of the style:

    ReturnType
    Name(Args)
    {
    }

Will now be styled as:

    ReturnType Name(Args) {

As always, if the signature is multi-line, the brace will be on the
following line.