Properly implement liveness analysis.
This was always a hack in the multi-parse compiler, but now, we can do
it properly. The live set is "all public functions" and from there we
calculate a reachability graph and mark what's live.

Unlike the old system, we do not bother changing read/written flags. We
use a new bit since that's easier to do when statics are split into
their own scope.

Now that liveness works correctly, the new compiler will no longer emit
code for unused functions.