Abandon the 1.11-dev branch.
The 1.11 branch still has serious bugs and they are essentially
unfixable given the massive divergence in the master branch. The old
code base is so unmaintainable that the burden is quite high to consider
fixing some of these problems.

Instead, let's abandon the 1.11-dev branch, and consider master to be
1.11 again. Although it is late in the cycle, I am confident enough in
the current codebase and its greatly improved logic and code generation.

Plugins compiled in the new compiler will not work on 1.10, due to the
new and improved array handling code. In addition, SourceMod will need
some small changes to smn_sorting.cpp before pinning to the new version.