Remove setjmp/longjmp.
setjmp and longjmp can be used as an old-school exception handling
method in C programs. In C++ they're rarely used since they do not
trigger destructors.

Removing these calls makes spcomp more embeddable. We can also take the
opportunity to clean up one of the weirder paths still remaining in the
compiler: fatal errors. These abort the compiler immediately. It's a
little tricky to mimic this without longjmp, and we cannot use C++
exceptions in our code. It's also tricky to propagate fatal errors out
of the lexer, since it can be called from pretty much anywhere.

Instead, we just remove the concept of a fatal error. The only scenario
it really helps with is when we're overwhelmed with errors, and we can
sort of retain this by periodically checking if we should stop parsing.