Build a big ol' binary.
It's time to start treating the SourcePawn tree as one homogenous
component, rather than separate mini-libraries. This step of the
refactoring process applies only to the build scripts.

The compiler and VM are now built as one component. They stay in
separate folders for organization, although, we will begin refactoring
that organization in the not too distant future.

Together with zlib and amtl, we produce a libsourcepawn archive. This is
used to create libsourcepawn.so (which is used by SourceMod).

Local tools (spcomp, spshell, and verifier) build using the archive.

This has the nice side effect of generally simplifying our build
scripts.

Support for building exp/* is now removed. Docgen will be re-added via
spcomp in a later patch.