Merge FunctionInfo and FunctionDecl.
The thinking behind the split was that we'd have a base class for
function expressions versus function statements. But it's too
cumbersome and wasteful.

The remaining grossness is that Decl::name is decorated, which
complicates code that only cares about the undecorated name. For now
I've resolved this by holding the undecorated name separately, but there
may be lingering error messages and stuff that uses the wrong one.