Refactor utf-8 parsing.
First and foremost, this refactors litchar() to be a bit more readable.
A lot of stuff has been dedented. It also now returns whether the
scanned value is a codepoint (and therefore must be converted to utf-8),
or whether it is a raw 8-bit value.

Since litchar could not previously distinguish these cases, it led to
some odd behavior. A single character literal could be a utf-8
codepoint, which makes no sense. Meanwhile a string literal was parsed
as a raw byte stream, which also makes no sense.

As if that wasn't enough, the utf-8 parser was bogus and did not work.
It is rewritten here and the new implementation should be much more
efficient as well.

The only functional change here is that double-quoted string literals
will now decode and re-encode utf-8 characters. This is intentional.
Scripts that were not properly utf-8 encoded will now emit warnings,
since there's no guarantee those strings will display properly for
clients.