Forbid out-of-range decimal escape sequences.
Pawn has an awful escape sequence that will interpret any number of
digits as a raw byte. Eg, "\04" and "\4" and "\000000000004" are the
same. The problem isn't the decimal part, it's that the number of
characters consumed is arbitrary. Eg, "\650" will be 650 truncated into
a byte, rather than "A0".

Unfortunately the cat is long out of the bag on this, and many plugins
rely on single or double digit decimal escape codes.

This change caps it to 3, and sets a maximum value of 255, which does
not break anything in the corpus.