Fix heap scopes not working with inline frames.
For performance reasons we do not create full C++ frames for each
function call, only for calls from C++ into scripts. Otherwise we use
"inline" frames which are much more lightweight. However, these inline
frames did not properly save and restore heap scopes like C++ frames
did.

Technically we could fix this by ensuring "heap.restore" is always
emittered at return points. It's nice though if the bytecode is as lean
as possible and the VM does any internal cleanup on its own.

For the interpreter we push the heap scope onto the heap, and restore it
when exiting inline frames.

For the JIT, we statically know at every exit point how many heap scopes
they are, and thus we can generate the precise exit code when we want
it.

Bug: issue #790
Test: new test case