Switch to a stream-based lexer.
The lexer was one of the few components not affected by the massive
rewrite of the parser in SourcePawn 1.11. It nonetheless has a bunch of
flaws that are finally being corrected.

The first is that the lexer attempted actual preprocessing. This would
make sense if Pawn had a preprocessing-only mode, but it no longer does,
and it's not clear why that would ever be a good idea anyway. This
preprocessing would replace comments with spaces, line continuations
with a bell character (which amusingly could lead to a sound effect
during compilation), and in-place substitution of macros. Since the
underlying character stream was affected, it was impossible to have
truly accurate line number and colum information.

The second flaw is that the lexer scanned line-by-line, which was slow
and fundamentally limiting. It greatly complicated the preprocessing
step.

The new lexer is not necessarily less or more complicated, but it is
certainly more readable, considerably faster, and much more powerful.
It now operates on the underlying file stream directly. This stream is
immutable, and the lexer scans it character by character. Comments are
skipped as needed. Preprocessor directives are handled as they are observed.
"Skipped" sections are handled by a special mini-preprocessor that only finds
"#else", "#elif", and "#endif" directives.

Most importantly, the egregiously complex "substallpatterns" function is
now gone. Macros are no longer expanded in place. Instead, a macro
introduces a new character stream just as an #include would.

This also formalizes how macro functions work. A macro function is of
the form:

    #define <identifier>(%1, %2, %3) blah(%1, %2, %3)

Arguments must be one of %0 through %9 and cannot be duplicated in the
argument list (but may be duplicated or not present at all in the
substituted expression).

The following patterns will no longer work:

    #define blah-
    #define blah();

These worked in the old lexer because pattern matched the entire macro
string, even including non-identifier characters. This is ridiculously
complex to implement efficiently, and macros are already discouraged, so
this behavior has been broken.

Additionally, a quirk in the old lexer prevented it from properly
erroring on patterns like this:

    #if defined X then

The "then" token was pushed but the "check if line empty" function used
the pending character stream, ignoring pushed tokens. This has been
fixed as well.