Remove symbol children.
Pawn's representation of "values" is original to the very first Small
compiler. It was previously unable to represent multi-dimensional
arrays. Rather than extend the value type, for some reason, the symbol
table was hacked instead. Every multi-dimensional array was actually a
series of nested symbols. This led to a lot of idiosyncratic and buggy
behavior, and greatly complicated a bunch of code. It also limited the
abilities of the language, since it was impossible to represent an array
type without a backing symbol.

This patch allows values to represent multi-dimensional arrays. This
improves correctness and simplifies a bunch of code. However it's still
a hack, and long-term, arrays should be represented as a type. This is
still a major step forward however.

Multi-dimensional arrays now declare only one symbol, and the "dim"
member of symbol is now gone.

This change will break some plugins that used const incorrectly. In the
old compiler, the nested symbols did not properly propagate const, so
only the "outer" dimension of the array was actually const. The new
behavior fixes this implicitly, and thus, plugins which mistakenly tried
to modify const arrays (or passed non-const arrays into const arguments)
will no longer compile.

Additionally, this patch introduces a new warning:

    const int X;

Will inform the user that const variables should be initialized.