Remove symbol::x.
This weird struct/union had an unused field. The remaining field has
been hoisted up and renamed "semantic_tag".

In addition, the meaning of this has been clarified. It is now set only
for "arrays" that have an implicit dimension (eg enum structs). The old
idiosyncratic uses around fields have been removed.

This is about a ~1% reduction in memory use.