Fix missing icon for spcomp.
This adds a product.version file and a generated header. This allows us
to generate sourcemod_version.h when not building SM.

The "#undef SM_USE_VERSIONLIB" is a huge hack. Really this should be
plumbed through the build scripts somehow, or SM could get rid of
versionlib which is not really useful anymore, or SM could hack a way
to opt SP out of versionlib. For now this will make SM build without
changing SM.

Bug: issue #867
Test: manual test on Windows