Remove support for function aliases.
This was an obscure feature used to make float.inc easier to write. It
was also briefly used to make methmodmaps easier to write. However it's
full of idiosyncracies and there is no type checking whatsoever for
aliases.

This will require a tiny bit of fixup on the SourceMod side, but
otherwise has no impact. The corpus will fail until SM is updated.