Remove symbol::usage.
This moves the read/write bits to VarDeclBase. FunctionDecl gets two new
bits: live (to indicate whether codegen is necessary) and maybe_used (to
indicate whether we should warn if the function is definitely not used).

This is a rather big change since "usage" had many references, so a
bunch of oddball places have been converted from symbol to Decl.