Remove Decl::addr.
This also fixes a long-standing, very obscure bug. Via default arguments
it is possible to reference a global variable before it's declared. When
"addr" was a wildly ambiguous field on "symbol", this went undetected
and the address defaulted to 0. When we moved it to Decl as an
std::optional, it exposed this bug.

By removing Decl::addr we can specialize addresses to each subclass. For
VarDecls, we can now use a Label instead, which automatically patches
code references as addresses resolve.

This also removes lots of silly duplication, eg Decl::addr and
FunctionDecl::CGInfo::label were the same thing.