Fix a bug allowing invalid array assignments.
For character arrays, we allow the left hand side to be bigger than the
rirght hand side of an array assignment. This makes sense for strings.
For other types, it was never supposed to be allowed, but for unclear
reasons (probably relating to old enum structs) was in fact valid.

This is now fixed. Three plugins were disabled in the corpus which had
invalid assignments.