Forbid spontaneous tag creation (aka, tag removal part 2).
Many years ago we introduced Transitional Syntax. Well, this tiny patch
is the start of the transition to just "syntax".

In the original Pawn language, tags were applied via labels, and labels
could appear anywhere. There was no formal method for creating them:
they simply existed as soon as they were used, and they behaved as if
there was an empty enum declared at global scope, even if the tag was
used locally.

This behavior is difficult to preserve, much less properly define, and
the globalness is particularly troublesome. By disabling this feature,
we're guaranteed that every type name references a type declaration in
the same scope chain.

This breaks 28 plugins in the corpus - a happily small number. These
plugins all have weird typos or are using labels in incomprehensible
ways. So, this esoteric language feature won't be missed.