Make enum structs into proper types.
When we first added enum structs to the compiler, it was pre-AST, and
the only way we could get it done was by tossing in tons and tons of
hacks to desugar enum structs into arrays. It required introducing the
concept of "semantic tags" that represented the intended type of an
array, since it would get desugared into cells, and very often we need
to de-desugar (sugar?) back into enum structs.

These hacks are particularly gross because they don't add any real value
to the compiler other than pain.

We've delayed fixing this for a very long time because, to date, Pawn
has been all about cells, and not at all about types. But with the AST,
and with `struct symbol` removed, and numerical tags removed, we can
finally cast off these shackles.

This is the first, official, truly non-cell type in Pawn. It's still got
plenty of hacks, but we've exposed all the nooks and crannies where
"variable size" equals "one cell", and that will prove very useful going
forward. This is also a starting point to eliminating iREFARRAY and
iARRAY which are pretty much complete nonsense.

This also means some changes are in order. A lot more semantic checks
need to make sure enum structs don't get bypassed. Calling matchtag()
and ignoring the result is no longer okay. We need additional checks to
find things matchtag() won't handle.

Anything that dealt with the size of an expression (eg, variable
declarations, sizeof, assignment) needs to check the underlying type
first, because it might be bigger than one cell.

Any kind of array code needed to be audited to see how it previously
interacted with enum structs.

Array and struct initialization continues to be handled by a common code
path since it's so similar.

Nested enum structs and enum structs with arrays had to be audited for
correctness.

sizeof() now takes a "hier1" expression, meaning a lot more stuff can
be tossed at it, and we can let the rest of the semantic pass calculate
sizes for us automatically.

There will be problems with this, and it'll take some time to root all
the corner cases out. But it's the right direction to go.