Remove iREFERENCE and add a new reference type.
Historically, Pawn constructed critical type information from two
sources of information: the tag, and the identifier kind. The identifier
kind describes whether the value refers to an array, variable,
expression result, reference, etc.

IdentifierKind is hugely problematic since it's not formally part of the
type. This means it can't participate in things like matchtag() and
find_userop(). It also creates situations where the two can become out
of sync. For example, changing iREFERENCE to iVARIABLE without an
RvalueExpr is a bug that is no longer possible.

With the recent removal of "semantic tags", we were able to remove
iENUMSTRUCT. Now, we are removing iREFERENCE, and creating a new
reference type. As before, this requires some auditing and additional
type checks. Mostly though, it is a simplification, since iVARIABLE and
iREFERENCE are equivalent in most code paths.